make targinfo[2] u_int8_t
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@21403 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/struct.h
@@ -219,7 +219,7 @@ struct LocalUser
 
 	/* target change stuff */
 	void *targets[10];		/* targets were aware of */
-	unsigned int targinfo[2];	/* cyclic array, no in use */
+	u_int8_t targinfo[2];	/* cyclic array, no in use */
 	time_t target_last;		/* last time we cleared a slot */
 };
